invoke from the menu
